[#23037] remove warning about int/uint comparison in CommandSyntax
Former-commit-id: eebae9a47fe0c21eeba43ac8e6f761b9b57403a6

+++ bibcxx/RunManager/CommandSyntax.h
@@ -331,25 +331,25 @@ class FactorKeyword
 
         bool isDoubleKeywordPresentInOccurence( string motCle, int num )
         {
-            if ( num >= _vectorOccurences.size() ) return false;
+            if ( (unsigned int)num >= _vectorOccurences.size() ) return false;
             return _vectorOccurences[num].isKeywordPresentDbl( motCle );
         };
 
         bool isIntegerKeywordPresentInOccurence( string motCle, int num )
         {
-            if ( num >= _vectorOccurences.size() ) return false;
+            if ( (unsigned int)num >= _vectorOccurences.size() ) return false;
             return _vectorOccurences[num].isKeywordPresentInt( motCle );
         };
 
         bool isStringKeywordPresentInOccurence( string motCle, int num )
         {
-            if ( num >= _vectorOccurences.size() ) return false;
+            if ( (unsigned int)num >= _vectorOccurences.size() ) return false;
             return _vectorOccurences[num].isKeywordPresentStr( motCle );
         };
 
         bool isOccurencePresent(int num)
         {
-            if ( num < _vectorOccurences.size() ) return true;
+            if ( (unsigned int)num < _vectorOccurences.size() ) return true;
             return false;
         };
